move map into its own little corner
