+++ frontends/mrxl/mrxl/gen_calyx.py
@@ -11,9 +11,11 @@ from calyx.py_ast import (
     While,
     ParComp,
     Control,
+    Empty,
 )
 from . import ast
 import calyx.builder as cb
+from . import map as map_impl
 
 
 class CompileError(Exception):
@@ -165,7 +167,8 @@ def gen_reduce_impl(
     return control
 
 
-def gen_map_impl(
+# ANCHOR: my_map_impl
+def my_map_impl(
     comp: cb.ComponentBuilder,
     dest: str,
     stmt: ast.Map,
@@ -184,92 +187,9 @@ def gen_map_impl(
       - a group that implements the loop condition, checking if the index
         has reached the end of the input array
     """
-
-    # Parallel loops representing the `map` body
-    map_loops = []
-
-    arr_size = arr_size // bank_factor
-    for bank in range(bank_factor):
-        suffix = f"b{bank}_{s_idx}"
-        idx = comp.reg(f"idx_{suffix}", 32)
-
-        # Increment the index
-        incr = incr_group(comp, idx, suffix)
-        # Check if we've reached the end of the loop
-        (port, cond) = cond_group(comp, idx, arr_size, suffix)
-
-        # Perform the computation
-        body = stmt.body
-        if isinstance(body, ast.LitExpr):  # Body is a constant
-            raise NotImplementedError()
-        if isinstance(body, ast.VarExpr):  # Body is a variable
-            raise NotImplementedError()
-
-        # Mapping from binding to arrays
-        name2arr = {bind.dst[0]: f"{bind.src}_b{bank}" for bind in stmt.binds}
-
-        def expr_to_port(expr: ast.BaseExpr):
-            if isinstance(expr, ast.LitExpr):
-                return cb.const(32, expr.value)
-            if isinstance(expr, ast.VarExpr):
-                return CompPort(CompVar(name2arr[expr.name]), "read_data")
-            raise CompileError(f"Unhandled expression: {type(expr)}")
-
-        # ANCHOR: map_op
-        if body.operation == "mul":
-            operation = comp.cell(
-                f"mul_{suffix}", Stdlib.op("mult_pipe", 32, signed=False)
-            )
-        else:
-            operation = comp.add(f"add_{suffix}", 32)
-        # ANCHOR_END: map_op
-
-        assert (
-            len(stmt.binds) <= 2
-        ), "Map statements with more than 2 arguments not supported"
-        # ANCHOR: map_inputs
-        with comp.group(f"eval_body_{suffix}") as evl:
-            # Index each array
-            for bind in stmt.binds:
-                # Map bindings have exactly one dest
-                mem = comp.get_cell(f"{name2arr[bind.dst[0]]}")
-                mem.addr0 = idx.out
-            # ANCHOR_END: map_inputs
-            # Provide inputs to the op
-            operation.left = expr_to_port(body.lhs)
-            operation.right = expr_to_port(body.rhs)
-            # ANCHOR: map_write
-            out_mem = comp.get_cell(f"{dest}_b{bank}")
-            out_mem.addr0 = idx.out
-            out_mem.write_data = operation.out
-            # Multipliers are sequential so we need to manipulate go/done signals
-            if body.operation == "mul":
-                operation.go = 1
-                out_mem.write_en = operation.done
-            else:
-                out_mem.write_en = 1
-            evl.done = out_mem.done
-            # ANCHOR_END: map_write
-
-        # Control to execute the groups
-        map_loops.append(
-            # ANCHOR: map_loop
-            While(
-                CompPort(CompVar(port), "out"),
-                CompVar(cond),
-                SeqComp(
-                    [
-                        Enable(f"eval_body_{suffix}"),
-                        Enable(incr),
-                    ]
-                ),
-            )
-            # ANCHOR_END: map_loop
-        )
-
-    control = ParComp(map_loops)
-
-    return control
+    # TODO: Implement map!
+    return Empty()
+    # ANCHOR_END: my_map_impl
 
 
 def gen_stmt_impl(
@@ -278,6 +198,7 @@ def gen_stmt_impl(
     arr_size: int,
     name2par: Dict[str, int],
     statement_idx: int,
+    use_my_map_impl: bool,
 ) -> Control:
     """
     Returns Calyx cells, wires, and control needed to implement
@@ -296,8 +217,17 @@ def gen_stmt_impl(
 
     name2par maps memory names to banking factors.
     """
-    if isinstance(stmt.operation, ast.Map):
-        return gen_map_impl(
+    if isinstance(stmt.operation, ast.Map) and use_my_map_impl:
+        return my_map_impl(
+            comp,
+            stmt.dst,
+            stmt.operation,
+            arr_size,
+            name2par[stmt.dst],
+            statement_idx,
+        )
+    elif isinstance(stmt.operation, ast.Map):
+        return map_impl.gen_map_impl(
             comp,
             stmt.dst,
             stmt.operation,
@@ -400,7 +330,7 @@ def reg_to_mem_group(
     return group_name
 
 
-def emit(prog: ast.Prog):
+def emit(prog: ast.Prog, use_my_map_impl: bool = False):
     """
     Returns a string containing a Calyx program, compiled from `prog`, a MrXL
     program.
@@ -467,7 +397,9 @@ def emit(prog: ast.Prog):
     control: List[Control] = []
     # Generate Calyx for each statement
     for i, stmt in enumerate(prog.stmts):
-        control.append(gen_stmt_impl(main, stmt, arr_size, par_factor, i))
+        control.append(
+            gen_stmt_impl(main, stmt, arr_size, par_factor, i, use_my_map_impl)
+        )
 
     # For each output register, move the value of the register into the external array
     if reg_to_mem:
